harit_model/requirements.txt
tensorflow
pandas
chainlit
kagglehub
openai
numpy
pillow
